Migrate last users of ForwardingTestCase and delete it.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=139349868
